+++ test/core/iomgr/tcp_posix_test.c
@@ -105,7 +105,7 @@ static size_t fill_socket_partial(int fd, size_t bytes) {
   do {
     write_bytes = write(fd, buf, bytes - total_bytes);
     if (write_bytes > 0) {
-      total_bytes += write_bytes;
+      total_bytes += (size_t)write_bytes;
     }
   } while ((write_bytes >= 0 || errno == EINTR) && bytes > total_bytes);
 
@@ -116,15 +116,15 @@ static size_t fill_socket_partial(int fd, size_t bytes) {
 
 struct read_socket_state {
   grpc_endpoint *ep;
-  ssize_t read_bytes;
-  ssize_t target_read_bytes;
+  size_t read_bytes;
+  size_t target_read_bytes;
   gpr_slice_buffer incoming;
   grpc_iomgr_closure read_cb;
 };
 
-static ssize_t count_slices(gpr_slice *slices, size_t nslices,
-                            int *current_data) {
-  ssize_t num_bytes = 0;
+static size_t count_slices(gpr_slice *slices, size_t nslices,
+                           int *current_data) {
+  size_t num_bytes = 0;
   unsigned i, j;
   unsigned char *buf;
   for (i = 0; i < nslices; ++i) {
@@ -140,7 +140,7 @@ static ssize_t count_slices(gpr_slice *slices, size_t nslices,
 
 static void read_cb(void *user_data, int success) {
   struct read_socket_state *state = (struct read_socket_state *)user_data;
-  ssize_t read_bytes;
+  size_t read_bytes;
   int current_data;
 
   GPR_ASSERT(success);
@@ -172,11 +172,11 @@ static void read_cb(void *user_data, int success) {
 }
 
 /* Write to a socket, then read from it using the grpc_tcp API. */
-static void read_test(ssize_t num_bytes, ssize_t slice_size) {
+static void read_test(size_t num_bytes, size_t slice_size) {
   int sv[2];
   grpc_endpoint *ep;
   struct read_socket_state state;
-  ssize_t written_bytes;
+  size_t written_bytes;
   gpr_timespec deadline = GRPC_TIMEOUT_SECONDS_TO_DEADLINE(20);
 
   gpr_log(GPR_INFO, "Read test of size %d, slice size %d", num_bytes,
@@ -222,7 +222,7 @@ static void read_test(ssize_t num_bytes, ssize_t slice_size) {
 
 /* Write to a socket until it fills up, then read from it using the grpc_tcp
    API. */
-static void large_read_test(ssize_t slice_size) {
+static void large_read_test(size_t slice_size) {
   int sv[2];
   grpc_endpoint *ep;
   struct read_socket_state state;
@@ -242,7 +242,7 @@ static void large_read_test(ssize_t slice_size) {
 
   state.ep = ep;
   state.read_bytes = 0;
-  state.target_read_bytes = written_bytes;
+  state.target_read_bytes = (size_t)written_bytes;
   gpr_slice_buffer_init(&state.incoming);
   grpc_iomgr_closure_init(&state.read_cb, read_cb, &state);
 
@@ -275,11 +275,11 @@ struct write_socket_state {
   int write_done;
 };
 
-static gpr_slice *allocate_blocks(ssize_t num_bytes, ssize_t slice_size,
+static gpr_slice *allocate_blocks(size_t num_bytes, size_t slice_size,
                                   size_t *num_blocks, int *current_data) {
-  size_t nslices = num_bytes / slice_size + (num_bytes % slice_size ? 1 : 0);
+  size_t nslices = num_bytes / slice_size + (num_bytes % slice_size ? 1u : 0u);
   gpr_slice *slices = gpr_malloc(sizeof(gpr_slice) * nslices);
-  ssize_t num_bytes_left = num_bytes;
+  size_t num_bytes_left = num_bytes;
   unsigned i, j;
   unsigned char *buf;
   *num_blocks = nslices;
@@ -334,7 +334,7 @@ void drain_socket_blocking(int fd, size_t num_bytes, size_t read_size) {
       GPR_ASSERT(buf[i] == current);
       current = (current + 1) % 256;
     }
-    bytes_left -= bytes_read;
+    bytes_left -= (size_t)bytes_read;
     if (bytes_left == 0) break;
   }
   flags = fcntl(fd, F_GETFL, 0);
@@ -366,7 +366,7 @@ static ssize_t drain_socket(int fd) {
 /* Write to a socket using the grpc_tcp API, then drain it directly.
    Note that if the write does not complete immediately we need to drain the
    socket in parallel with the read. */
-static void write_test(ssize_t num_bytes, ssize_t slice_size) {
+static void write_test(size_t num_bytes, size_t slice_size) {
   int sv[2];
   grpc_endpoint *ep;
   struct write_socket_state state;
@@ -400,7 +400,7 @@ static void write_test(ssize_t num_bytes, ssize_t slice_size) {
     case GRPC_ENDPOINT_DONE:
       /* Write completed immediately */
       read_bytes = drain_socket(sv[0]);
-      GPR_ASSERT(read_bytes == num_bytes);
+      GPR_ASSERT((size_t)read_bytes == num_bytes);
       break;
     case GRPC_ENDPOINT_PENDING:
       drain_socket_blocking(sv[0], num_bytes, num_bytes);
@@ -426,7 +426,7 @@ static void write_test(ssize_t num_bytes, ssize_t slice_size) {
 }
 
 void run_tests(void) {
-  int i = 0;
+  size_t i = 0;
 
   read_test(100, 8192);
   read_test(10000, 8192);
